+++ src/egl/opengl/eglgears.c
@@ -220,7 +220,7 @@ idle(void)
   t0 = t;
 
   angle += 70.0 * dt;  /* 70 degrees per second */
-  angle = fmod(angle, 360.0); /* prevents eventual overflow */
+  angle = fmodf(angle, 360.0f); /* prevents eventual overflow */
 
   eglutPostRedisplay();
 }
