+++ src/egl/opengles1/gears.c
@@ -351,7 +351,7 @@ gears_idle(void)
   t0 = t;
 
   angle += 70.0 * dt;  /* 70 degrees per second */
-  angle = fmod(angle, 360.0); /* prevents eventual overflow */
+  angle = fmodf(angle, 360.0f); /* prevents eventual overflow */
 
   eglutPostRedisplay();
 }
